Change hit ratio display format

+++ app/src/main/java/com/slotnslot/slotnslot/models/SlotRoom.java
@@ -38,7 +38,7 @@ public class SlotRoom implements Serializable {
         this.bankerBalance = bankerBalance;
     }
 
-    public int getHitRatio() {
-        return (int) (hitRatio * 100);
+    public double getHitRatio() {
+        return hitRatio * 100;
     }
 }
